third_party/blink/web_tests/external/wpt/service-workers/idlharness.https.any.sharedworker-expected.txt

This is a testharness.js-based test.
Found 40 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] ServiceWorker interface: existence and properties of interface object
  assert_own_property: self does not have own property "ServiceWorker" expected property "ServiceWorker" missing
[FAIL] ServiceWorker interface object length
  assert_own_property: self does not have own property "ServiceWorker" expected property "ServiceWorker" missing
[FAIL] ServiceWorker interface object name
  assert_own_property: self does not have own property "ServiceWorker" expected property "ServiceWorker" missing
[FAIL] ServiceWorker interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "ServiceWorker" expected property "ServiceWorker" missing
[FAIL] ServiceWorker interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "ServiceWorker" expected property "ServiceWorker" missing
[FAIL] ServiceWorker interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "ServiceWorker" expected property "ServiceWorker" missing
[FAIL] ServiceWorker interface: attribute scriptURL
  assert_own_property: self does not have own property "ServiceWorker" expected property "ServiceWorker" missing
[FAIL] ServiceWorker interface: attribute state
  assert_own_property: self does not have own property "ServiceWorker" expected property "ServiceWorker" missing
[FAIL] ServiceWorker interface: operation postMessage(any, sequence<object>)
  assert_own_property: self does not have own property "ServiceWorker" expected property "ServiceWorker" missing
[FAIL] ServiceWorker interface: operation postMessage(any, optional StructuredSerializeOptions)
  assert_own_property: self does not have own property "ServiceWorker" expected property "ServiceWorker" missing
[FAIL] ServiceWorker interface: attribute onstatechange
  assert_own_property: self does not have own property "ServiceWorker" expected property "ServiceWorker" missing
[FAIL] ServiceWorkerContainer interface: existence and properties of interface object
  assert_own_property: self does not have own property "ServiceWorkerContainer" expected property "ServiceWorkerContainer" missing
[FAIL] ServiceWorkerContainer interface object length
  assert_own_property: self does not have own property "ServiceWorkerContainer" expected property "ServiceWorkerContainer" missing
[FAIL] ServiceWorkerContainer interface object name
  assert_own_property: self does not have own property "ServiceWorkerContainer" expected property "ServiceWorkerContainer" missing
[FAIL] ServiceWorkerContainer interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "ServiceWorkerContainer" expected property "ServiceWorkerContainer" missing
[FAIL] ServiceWorkerContainer interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "ServiceWorkerContainer" expected property "ServiceWorkerContainer" missing
[FAIL] ServiceWorkerContainer interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "ServiceWorkerContainer" expected property "ServiceWorkerContainer" missing
[FAIL] ServiceWorkerContainer interface: attribute controller
  assert_own_property: self does not have own property "ServiceWorkerContainer" expected property "ServiceWorkerContainer" missing
[FAIL] ServiceWorkerContainer interface: attribute ready
  assert_own_property: self does not have own property "ServiceWorkerContainer" expected property "ServiceWorkerContainer" missing
[FAIL] ServiceWorkerContainer interface: operation register((TrustedScriptURL or USVString), optional RegistrationOptions)
  assert_own_property: self does not have own property "ServiceWorkerContainer" expected property "ServiceWorkerContainer" missing
[FAIL] ServiceWorkerContainer interface: operation getRegistration(optional USVString)
  assert_own_property: self does not have own property "ServiceWorkerContainer" expected property "ServiceWorkerContainer" missing
[FAIL] ServiceWorkerContainer interface: operation getRegistrations()
  assert_own_property: self does not have own property "ServiceWorkerContainer" expected property "ServiceWorkerContainer" missing
[FAIL] ServiceWorkerContainer interface: operation startMessages()
  assert_own_property: self does not have own property "ServiceWorkerContainer" expected property "ServiceWorkerContainer" missing
[FAIL] ServiceWorkerContainer interface: attribute oncontrollerchange
  assert_own_property: self does not have own property "ServiceWorkerContainer" expected property "ServiceWorkerContainer" missing
[FAIL] ServiceWorkerContainer interface: attribute onmessage
  assert_own_property: self does not have own property "ServiceWorkerContainer" expected property "ServiceWorkerContainer" missing
[FAIL] ServiceWorkerContainer interface: attribute onmessageerror
  assert_own_property: self does not have own property "ServiceWorkerContainer" expected property "ServiceWorkerContainer" missing
[FAIL] ServiceWorkerContainer must be primary interface of navigator.serviceWorker
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Stringification of navigator.serviceWorker
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] ServiceWorkerContainer interface: navigator.serviceWorker must inherit property "controller" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] ServiceWorkerContainer interface: navigator.serviceWorker must inherit property "ready" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] ServiceWorkerContainer interface: navigator.serviceWorker must inherit property "register((TrustedScriptURL or USVString), optional RegistrationOptions)" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] ServiceWorkerContainer interface: calling register((TrustedScriptURL or USVString), optional RegistrationOptions) on navigator.serviceWorker with too few arguments must throw TypeError
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] ServiceWorkerContainer interface: navigator.serviceWorker must inherit property "getRegistration(optional USVString)" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] ServiceWorkerContainer interface: calling getRegistration(optional USVString) on navigator.serviceWorker with too few arguments must throw TypeError
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] ServiceWorkerContainer interface: navigator.serviceWorker must inherit property "getRegistrations()" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] ServiceWorkerContainer interface: navigator.serviceWorker must inherit property "startMessages()" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] ServiceWorkerContainer interface: navigator.serviceWorker must inherit property "oncontrollerchange" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] ServiceWorkerContainer interface: navigator.serviceWorker must inherit property "onmessage" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] ServiceWorkerContainer interface: navigator.serviceWorker must inherit property "onmessageerror" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] WorkerNavigator interface: attribute serviceWorker
  assert_true: The prototype object must have a property "serviceWorker" expected true got false
Harness: the test ran to completion.
